refactor: reorganize NewSettings module structure to separate concerns
