+++ src/views/NewSettings/components/Sections/Account/Account/AccountSection.tsx
@@ -6,7 +6,7 @@ import { useSelector } from 'react-redux';
 import { areCredentialsCorrect } from 'app/auth/services/auth.service';
 import userService from 'app/auth/services/user.service';
 import errorService from 'app/core/services/error.service';
-import ClearTrashDialog from 'views/Trash/components/ClearTrashDialog';
+import ClearTrashDialog from '../../../../../../views/Trash/components/ClearTrashDialog';
 import { useTranslationContext } from 'app/i18n/provider/TranslationProvider';
 import notificationsService, { ToastType } from 'app/notifications/services/notifications.service';
 import { Button } from '@internxt/ui';
