+++ src/views/NewSettings/utils/planUtils.ts
@@ -9,7 +9,7 @@ import { bytesToString } from 'app/drive/services/size.service';
 import { t } from 'i18next';
 import { FreeStoragePlan } from 'app/drive/types';
 import moneyService from 'app/payment/services/currency.service';
-import { ChangePlanType } from '../components/PlanCard';
+import { ChangePlanType } from '../components/Sections/Account/Plans/components/PlanCard';
 
 function displayAmount(value: number, decimalPoints = 2) {
   return (value / 100).toFixed(decimalPoints);
